feat(front): mv toSearchHistory in a new adapter file
